WIP: cancellation of subscription
